+++ lib/view/screens/customize_widgets_screen.dart
@@ -5,14 +5,12 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_blurhash/flutter_blurhash.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:provider/provider.dart' as provider;
 import 'package:workout_player/models/combined/auth_and_database.dart';
 import 'package:workout_player/models/user.dart';
 import 'package:workout_player/generated/l10n.dart';
-import 'package:workout_player/services/auth.dart';
-import 'package:workout_player/services/database.dart';
 import 'package:workout_player/styles/text_styles.dart';
 import 'package:workout_player/view/widgets/buttons.dart';
+import 'package:workout_player/view/widgets/widgets.dart';
 import 'package:workout_player/view_models/customize_widgets_screen_model.dart';
 import 'package:workout_player/view_models/progress_tab_model.dart';
 
@@ -32,18 +30,14 @@ class CustomizeWidgetsScreen extends StatefulWidget {
     BuildContext context, {
     required User user,
   }) {
-    final auth = provider.Provider.of<AuthBase>(context, listen: false);
-    final database = provider.Provider.of<Database>(context, listen: false);
-
-    Navigator.of(context, rootNavigator: true).push(
-      CupertinoPageRoute(
-        fullscreenDialog: true,
-        builder: (context) => Consumer(
-          builder: (context, watch, child) => CustomizeWidgetsScreen(
-            user: user,
-            model: watch(customizeWidgetsScreenModelProvider),
-            authAndDatabase: AuthAndDatabase(auth: auth, database: database),
-          ),
+    customPush(
+      context,
+      rootNavigator: true,
+      builder: (context, auth, database) => Consumer(
+        builder: (context, watch, child) => CustomizeWidgetsScreen(
+          user: user,
+          model: watch(customizeWidgetsScreenModelProvider),
+          authAndDatabase: AuthAndDatabase(auth: auth, database: database),
         ),
       ),
     );