+++ lib/view/screens/home_screen.dart
@@ -18,10 +18,10 @@ class HomeScreen extends StatefulWidget {
 
 class _HomeScreenState extends State<HomeScreen> {
   @override
-  void initState() {
-    super.initState();
+  void didChangeDependencies() {
+    super.didChangeDependencies();
     context.read(homeScreenModelProvider).updateUser(context);
-    context.read(homeScreenModelProvider).setMiniplayerHeight();
+    context.read(homeScreenModelProvider).setMiniplayerHeight(context);
   }
 
   @override