+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Hi, below is all the changes from my project, ~~Herakles~~ Break, after version 
 
 ### Bug Fix & Refactor
 - Fix Android's back button issue
+- Fix Miniplayer's size issue for all devices
 
 ## [0.3.6] - October 8th, 2021
 ### New Features & Screens